+++ api/api/routes/achievements.py
@@ -1,0 +1,21 @@
+from flask import Blueprint
+import api
+
+from api.common import WebSuccess, WebError
+from api.annotations import api_wrapper, require_login
+from api.annotations import block_before_competition, block_after_competition
+from api.annotations import log_action
+
+blueprint = Blueprint("achievements_api", __name__, template_folder="templates")
+
+@blueprint.route('', methods=['GET'])
+@require_login
+@api_wrapper
+def get_achievements_hook():
+    tid = api.user.get_team()["tid"]
+    achievements = api.achievement.get_earned_achievements_display(tid=tid)
+
+    for achievement in achievements:
+        achievement["timestamp"] = None  # JB : Hack to temporarily fix achievements timestamp problem
+
+    return WebSuccess(data=achievements)
